+++ include/bit_sequence.hpp
@@ -24,12 +24,12 @@ class BitSequence : public Sequence<Bit> {
     }
   }
 
-  BitSequence(const Bit* items, int count) : words_(), length_(0) {
+  BitSequence(const Bit* items, size_t count) : words_(), length_(0) {
     if (count < 0) {
       throw InvalidArgument("Ошибка, отрицательное количество битов");
     }
     ResizeBits(count);
-    for (int i = 0; i < count; ++i) {
+    for (size_t i = 0; i < count; ++i) {
       SetBitInternal(i, items[i].Value() != 0);
     }
   }
@@ -51,12 +51,12 @@ class BitSequence : public Sequence<Bit> {
     return Get(length_ - 1);
   }
 
-  Bit Get(int index) const override {
+  Bit Get(size_t index) const override {
     ValidateIndex(index);
     return Bit(GetBitInternal(index));
   }
 
-  int GetLength() const override {
+  size_t GetLength() const override {
     return length_;
   }
 
@@ -68,11 +68,11 @@ class BitSequence : public Sequence<Bit> {
     return *this;
   }
 
-  Sequence<Bit>* GetSubsequence(int start_index, int end_index) const override {
+  Sequence<Bit>* GetSubsequence(size_t start_index, size_t end_index) const override {
     ValidateClosedRange(start_index, end_index);
     BitSequence* result = new BitSequence();
     result->ResizeBits(end_index - start_index + 1);
-    for (int i = start_index; i <= end_index; ++i) {
+    for (size_t i = start_index; i <= end_index; ++i) {
       result->SetBitInternal(i - start_index, GetBitInternal(i));
     }
     return result;
@@ -88,36 +88,36 @@ class BitSequence : public Sequence<Bit> {
     return InsertAt(item, 0, true);
   }
 
-  Sequence<Bit>* InsertAt(const Bit& item, int index) override {
+  Sequence<Bit>* InsertAt(const Bit& item, size_t index) override {
     if (index < 0 || index >= length_) {
       throw IndexOutOfRange("Ошибка, индекс не из диапазона");
     }
     return InsertAt(item, index, false);
   }
 
-  Sequence<Bit>* Set(int index, const Bit& item) override {
+  Sequence<Bit>* Set(size_t index, const Bit& item) override {
     ValidateIndex(index);
     SetBitInternal(index, item.Value() != 0);
     return this;
   }
 
-  Sequence<Bit>* Slice(int index, int count, const Sequence<Bit>* replacement = nullptr) override {
+  Sequence<Bit>* Slice(size_t index, size_t count, const Sequence<Bit>* replacement = nullptr) override {
     if (count < 0) {
       throw InvalidArgument("Ошибка, отрицательное количество индексов, которые нужно удалить");
     }
-    const int start = NormalizeSliceIndex(index);
-    const int available = length_ - start;
-    const int remove_count = (count < available) ? count : available;
+    const size_t start = NormalizeSliceIndex(index);
+    const size_t available = length_ - start;
+    const size_t remove_count = (count < available) ? count : available;
     BitSequence rebuilt;
-    for (int i = 0; i < start; ++i) {
+    for (size_t i = 0; i < start; ++i) {
       rebuilt.Append(Get(i));
     }
     if (replacement != nullptr) {
-      for (int i = 0; i < replacement->GetLength(); ++i) {
+      for (size_t i = 0; i < replacement->GetLength(); ++i) {
         rebuilt.Append(replacement->Get(i));
       }
     }
-    for (int i = start + remove_count; i < length_; ++i) {
+    for (size_t i = start + remove_count; i < length_; ++i) {
       rebuilt.Append(Get(i));
     }
     *this = rebuilt;
@@ -125,7 +125,7 @@ class BitSequence : public Sequence<Bit> {
   }
 
   Sequence<Bit>* Concat(const Sequence<Bit>& other) override {
-    for (int i = 0; i < other.GetLength(); ++i) {
+    for (size_t i = 0; i < other.GetLength(); ++i) {
       Append(other.Get(i));
     }
     return this;
@@ -146,7 +146,7 @@ class BitSequence : public Sequence<Bit> {
   BitSequence* And(const BitSequence& other) const {
     ValidateEqualLength(other);
     BitSequence* result = new BitSequence(*this);
-    for (int i = 0; i < WordCount(); ++i) {
+    for (size_t i = 0; i < WordCount(); ++i) {
       result->words_.Set(i, words_.Get(i) & other.words_.Get(i));
     }
     result->ClearTailBits();
@@ -156,7 +156,7 @@ class BitSequence : public Sequence<Bit> {
   BitSequence* Or(const BitSequence& other) const {
     ValidateEqualLength(other);
     BitSequence* result = new BitSequence(*this);
-    for (int i = 0; i < WordCount(); ++i) {
+    for (size_t i = 0; i < WordCount(); ++i) {
       result->words_.Set(i, words_.Get(i) | other.words_.Get(i));
     }
     result->ClearTailBits();
@@ -166,7 +166,7 @@ class BitSequence : public Sequence<Bit> {
   BitSequence* Xor(const BitSequence& other) const {
     ValidateEqualLength(other);
     BitSequence* result = new BitSequence(*this);
-    for (int i = 0; i < WordCount(); ++i) {
+    for (size_t i = 0; i < WordCount(); ++i) {
       result->words_.Set(i, words_.Get(i) ^ other.words_.Get(i));
     }
     result->ClearTailBits();
@@ -175,7 +175,7 @@ class BitSequence : public Sequence<Bit> {
 
   BitSequence* Not() const {
     BitSequence* result = new BitSequence(*this);
-    for (int i = 0; i < WordCount(); ++i) {
+    for (size_t i = 0; i < WordCount(); ++i) {
       result->words_.Set(i, ~words_.Get(i));
     }
     result->ClearTailBits();
@@ -185,7 +185,7 @@ class BitSequence : public Sequence<Bit> {
   std::string ToBitString() const {
     std::string result;
     result.reserve(length_);
-    for (int i = 0; i < length_; ++i) {
+    for (size_t i = 0; i < length_; ++i) {
       result.push_back(GetBitInternal(i) ? '1' : '0');
     }
     return result;
@@ -193,22 +193,22 @@ class BitSequence : public Sequence<Bit> {
 
  private:
   DynamicArray<std::uint64_t> words_;
-  int length_;
+  size_t length_;
   
-  int WordCount() const {
+  size_t WordCount() const {
     return (length_ + 63) / 64;
   }
 
-  static int BitOffset(int bit_index) {
+  static size_t BitOffset(size_t bit_index) {
     return bit_index % 64;
   }
 
-  void ResizeBits(int new_length) {
+  void ResizeBits(size_t new_length) {
     if (new_length < 0) {
       throw InvalidArgument("Ошибка, отрицательная длина");
     }
-    const int old_word_count = WordCount();
-    const int new_word_count = (new_length + 63) / 64;
+    const size_t old_word_count = WordCount();
+    const size_t new_word_count = (new_length + 63) / 64;
     if (new_word_count != old_word_count) {
       words_.Resize(new_word_count);
     }
@@ -216,12 +216,12 @@ class BitSequence : public Sequence<Bit> {
     ClearTailBits();
   }
 
-  bool GetBitInternal(int index) const {
+  bool GetBitInternal(size_t index) const {
     const std::uint64_t word = words_.Get(WordIndex(index));
     return ((word >> BitOffset(index)) & 1ULL) != 0ULL;
   }
 
-  void SetBitInternal(int index, bool value) {
+  void SetBitInternal(size_t index, bool value) {
     std::uint64_t word = words_.Get(WordIndex(index));
     const std::uint64_t mask = 1ULL << BitOffset(index);
     if (value) {
@@ -236,7 +236,7 @@ class BitSequence : public Sequence<Bit> {
     if (length_ == 0 || WordCount() == 0) {
       return;
     }
-    const int used_bits = length_ % 64;
+    const size_t used_bits = length_ % 64;
     if (used_bits == 0) {
       return;
     }
@@ -244,13 +244,13 @@ class BitSequence : public Sequence<Bit> {
     words_.Set(WordCount() - 1, words_.Get(WordCount() - 1) & mask);
   }
 
-  void ValidateIndex(int index) const {
+  void ValidateIndex(size_t index) const {
     if (index < 0 || index >= length_) {
       throw IndexOutOfRange("Ошибка, индекс не из диапазона");
     }
   }
 
-  void ValidateClosedRange(int start_index, int end_index) const {
+  void ValidateClosedRange(size_t start_index, size_t end_index) const {
     if (length_ == 0) {
       throw EmptyCollection("Ошибка, длина равна 0");
     }
@@ -262,8 +262,8 @@ class BitSequence : public Sequence<Bit> {
     }
   }
 
-  int NormalizeSliceIndex(int index) const {
-    int normalized = index;
+  size_t NormalizeSliceIndex(size_t index) const {
+    size_t normalized = index;
     if (normalized < 0) {
       normalized = length_ + normalized;
     }
@@ -279,11 +279,11 @@ class BitSequence : public Sequence<Bit> {
     }
   }
 
-  static int WordIndex(int bit_index) {
+  static size_t WordIndex(size_t bit_index) {
     return bit_index / 64;
   }
 
-  Sequence<Bit>* InsertAt(const Bit& item, int index, bool allow) {
+  Sequence<Bit>* InsertAt(const Bit& item, size_t index, bool allow) {
     if (!allow && length_ == 0) {
       throw IndexOutOfRange("Ошибка, пустая последовательность");
     }
@@ -295,9 +295,9 @@ class BitSequence : public Sequence<Bit> {
     if (index < 0 || index > length_ || (!allow && index == length_)) {
       throw IndexOutOfRange("Ошибка, индекс не из диапазона");
     }
-    const int old_length = length_;
+    const size_t old_length = length_;
     ResizeBits(length_ + 1);
-    for (int i = old_length; i > index; --i) {
+    for (size_t i = old_length; i > index; --i) {
       SetBitInternal(i, GetBitInternal(i - 1));
     }
     SetBitInternal(index, item.Value() != 0);
